+++ methods/src/test/kotlin/methods/expand/ExpandRulesTest.kt
@@ -1,5 +1,7 @@
 package methods.expand
 
+import engine.context.BooleanSetting
+import engine.context.Setting
 import engine.methods.testMethod
 import engine.methods.testRule
 import methods.expand.ExpandRules.ApplyFoilMethod
@@ -166,6 +168,35 @@ class ExpandRulesTest {
             "x*(1 + sqrt[3])",
             DistributeMultiplicationOverSum,
             null,
+        )
+    }
+
+    @Test
+    fun `test ExpandProductOfSingleTermAndSum with CopySumSignsWhenDistributing`() {
+        val settings = mapOf(Setting.CopySumSignsWhenDistributing to BooleanSetting.True)
+        testRule(
+            "5(x - 2)",
+            DistributeMultiplicationOverSum,
+            "5*x - 5*2",
+            settings = settings,
+        )
+        testRule(
+            "5(-2x + 2)",
+            DistributeMultiplicationOverSum,
+            "-5*<. 2x .> + 5*2",
+            settings = settings,
+        )
+        testRule(
+            "(-5)(-2x - 2 + (-4))",
+            DistributeMultiplicationOverSum,
+            "-(-5)*<. 2x .> - (-5)*2 + (-5)*(-4)",
+            settings = settings,
+        )
+        testRule(
+            "-5(3x - 2)",
+            DistributeMultiplicationOverSum,
+            "(-5)*<. 3x .> - (-5)*2",
+            settings = settings,
         )
     }